Commit of work prior to departure for EPS. This appears to have a working fieldtest and getfield, though getfield is limited to 3-D by the hack fix.

+++ makefile
@@ -2,12 +2,14 @@ LIBRARIES = -L/usr/X11R6/lib/ -L/home/hutch/accis/ -laccisX -lXt -lX11
 # Things just needed for the test routine:
 UTILITIES=udisplay.o
 # The sormpi system.
-OBJECTS=sormpi.o sorrelaxgen.o mpibbdy.o  cijroutine.o cijplot.o 3dobjects.o mditerate.o interpolations.o svdsol.o
+OBJECTS=sormpi.o sorrelaxgen.o mpibbdy.o  cijroutine.o cijplot.o 3dobjects.o mditerate.o interpolations.o svdsol.o getfield.o
 HEADERS=bbdydecl.f sormesh.f objcom.f 3dcom.f
 TARGETS=mpibbdytest mditeratetest sormpitest
 G77=mpif77
 #COMPILE-SWITCHES = -Wall -O2  -I. 
 COMPILE-SWITCHES = -Wall  -I. -g  -ffortran-bounds-check
+
+# Things to compile without the standard switches
 
 
 % : %.f makefile $(OBJECTS) $(UTILITIES)
@@ -18,12 +20,18 @@ COMPILE-SWITCHES = -Wall  -I. -g  -ffortran-bounds-check
 
 #default target
 smt.out : sormpitest
-	mv smt.out smt.prev
+	if [ -f smt.out ] ; then mv smt.out smt.prev ; fi
 #	mpiexec -n 8 ./sormpitest -p
 #	mpiexec -n 1 ./sormpitest -p
 	./sormpitest -p
 #	diff testing/smt.fixed smt.out
 	diff smt.prev smt.out
+
+interpolations.o : interpolations.f makefile $(HEADERS)
+	$(G77)  -c -Wall $*.f
+
+getfield.o : getfield.f makefile $(HEADERS)
+	$(G77)  -c -Wall $*.f
 
 sormpitest : sormpitest.f makefile $(OBJECTS) $(UTILITIES)
 	$(G77)  -o sormpitest $(COMPILE-SWITCHES) sormpitest.f  $(OBJECTS) $(UTILITIES) $(LIBRARIES)
